third_party/WebKit/LayoutTests/fast/dom/HTMLObjectElement/HTMLObject-contentWindow-query.html

<!DOCTYPE html>
<title>HTMLObjectElement.contentWindow with SVG document</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<object data="resources/HTMLObject-contentWindow.svg"></object>
<script>
  async_test(function(t) {
    var obj = document.querySelector('object');
    obj.onload = function () {
        t.step(function() {
          assert_equals(obj.contentWindow.attrInContentWindow, "validAttr");
        });
        t.done()
    }
  }, "Test that content window and attributes should be accessible.");
</script>
</body>
